fix: rabbitai warnings, only cd to executable path on win
